+++ src/app/core/tools/WiringTool.ts
@@ -2,7 +2,7 @@ import {Vector} from "Vector";
 
 import {Event}       from "core/utils/Events";
 import {CircuitInfo} from "core/utils/CircuitInfo";
-import {GetAllPorts} from "core/utils/ComponentUtils";
+import {FindPorts, GetAllPorts} from "core/utils/ComponentUtils";
 
 import {ConnectionAction} from "core/actions/addition/ConnectionAction";
 
@@ -20,43 +20,6 @@ export const WiringTool = (() => {
     let wire: Wire;
     let stateType: StateType;
 
-    function findPorts({input, camera, designer}: Partial<CircuitInfo>): Port[] {
-        const worldMousePos = camera.getWorldPos(input.getMousePos());
-        const objects = designer.getObjects().reverse();
-
-        const noSObj = objects.filter(o => o.isWithinSelectBounds(worldMousePos)).length == 0;
-        let pressables: Pressable[] = [];
-        for (let i = 0; i < objects.length; ++i) {
-            if (isPressable(objects[i]))
-                pressables.push(objects[i] as Pressable);
-        }
-        const noPObj = pressables.filter(o => o.isWithinPressBounds(worldMousePos)).length == 0;
-        if (noSObj && noPObj)
-            // Look through all ports in all objects
-            //  and find one where the mouse is over
-            return GetAllPorts(objects).filter(p => p.isWithinSelectBounds(worldMousePos));
-
-        // https://github.com/OpenCircuits/OpenCircuits/issues/624
-        let output: Port[] = [];
-        let notVisible: Port[] = [];
-        for (let i = 0; i < objects.length; ++i) {
-            const sObj = objects[i].isWithinSelectBounds(worldMousePos);
-            const pObj = (isPressable(objects[i]) && (objects[i] as Pressable).isWithinPressBounds(worldMousePos));
-            if (sObj || pObj) {
-                // selectable ports
-                let ports = GetAllPorts(objects).filter(p => p.isWithinSelectBounds(worldMousePos));
-                for (let j = 0; j < ports.length; ++j) {
-                    let dadIndex = objects.findIndex(o => o === ports[j].getParent());
-                    if (dadIndex != -1 && dadIndex <= i) output.push(ports[j]);
-                    else notVisible.push(ports[j]);
-                }
-            }
-        }
-        // Ensure ports are not in middle of stacked objects
-        return output.filter(function(o) {
-            return notVisible.indexOf(o) == -1;
-        });
-    }
     function setWirePoint(v: Vector): void {
         // The wiring tool always starts with 1 port connected
         //  and the other point should be following the mouse
@@ -78,7 +41,7 @@ export const WiringTool = (() => {
             const {locked, input, camera, designer} = info;
             if (locked)
                 return false;
-            const ports = findPorts(info);
+            const ports = FindPorts(info);
             // Activate if the user drags or clicks on a port
             return ((event.type === "mousedown" && input.getTouchCount() === 1) ||
                     (event.type === "click")) &&
@@ -97,7 +60,7 @@ export const WiringTool = (() => {
 
 
         onActivate(event: Event, info: CircuitInfo): void {
-            port = findPorts(info)[0];
+            port = FindPorts(info)[0];
 
             // Create wire and set it's other point to be at `port`
             wire = info.designer.createWire(port, undefined);
@@ -108,7 +71,7 @@ export const WiringTool = (() => {
         onDeactivate({}: Event, info: CircuitInfo): void {
             const {history, designer} = info;
             // See if we ended on a port
-            const port2 = findPorts(info).find(p => wire.canConnectTo(p));
+            const port2 = FindPorts(info).find(p => wire.canConnectTo(p));
             if (port2 !== undefined)
                 history.add(new ConnectionAction(designer, port, port2).execute());
         },
@@ -125,24 +88,6 @@ export const WiringTool = (() => {
 
         getWire(): Wire {
             return wire;
-        },
-
-        /**
-         * Checks for ports over objects using findPorts
-         * @param info CircuitInfo to be used by findPorts
-         * @returns whether there are any visible ports
-         */
-        visiblePorts(info: CircuitInfo): boolean {
-            return findPorts(info).length > 0;
-        },
-
-        /**
-         * To test for special case Node ports
-         * @param info CircuitInfo to be used by findPorts
-         * @returns whether there are visible ports belonging to a Node
-         */
-        nodePorts(info: CircuitInfo): boolean {
-            return findPorts(info).some(o => isNode(o.getParent()));
         }
     }
 })();
